adding .gitignore, testng.xml; renaming packages
