+++ src/test/resources/testng.xml
@@ -1,0 +1,9 @@
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+
+<suite name="Suite1" verbose="1" >
+    <test name="Regression1"   >
+        <packages>
+            <package name="test.java.com.jeff.tests" />
+        </packages>
+    </test>
+</suite>
